+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* added credentials authentication
 * added CI test
 
 ## 1.5.1